\name{ppSplines}
\alias{ppSplines}
\title{
Spline design matrix.
}
\description{
The function builds a matrix for piecewise polynomials fit. 
}
\usage{
ppSplines(form, data, knots = NA, pFact = 0.3,
          deg = 1, method = c("manual", "bs"), minZeroProp=0.1,
          subGroup=NULL)
}
\arguments{
  \item{form}{A formula to create the vectors to be converted into
    spline matrices. See the example below}
  \item{data}{A \code{data.frame} that contains all variables from the
    formula \code{form}}
  \item{knots}{The piecewise polynomial knots. If set to \code{NA}, the
    knots are set to the \code{p} equally spaced quantiles of \code{X}. If
    \code{NULL}, the function returns \code{X}.}
  \item{pFact}{The number of knots when the argument \code{knots} is set to
    \code{NA} if \code{n^pFact}, where n is the length of \code{X}.}
  \item{deg}{The degree of the piecewise polynomials.}
  \item{method}{Should the method be homemade (manual) of based on the
    \code{bs} function from the splines package?}
  \item{minZeroProp}{If the proportion of zeros is greater than this
    value, the knots and spline matrices are based on the non-zero
    observations. This is particularly useful when binary variables are
    interacted with continuous variables.}
  \item{subGroup}{A binary or logical vector. If provided the basis
    matrix is built only for observations with \code{subGroup} either
    \code{TRUE} or equal to 1. The other rows in the spline matrix are
    set to zero. }
}

\examples{
data(simData)
all <- ppSplines(~X1*X2, simData)
## The following shows the knots
## for each variable generated by the formula
attr(all, "knots")
}

\keyword{causal effects}
