\name{simData}
\alias{simData}
\docType{data}
\title{
Simulated data.
}
\description{
This dataset is used in several documentation files to illustrate the
different functionalities of the package.
}
\usage{data("simData")}
\format{
  A data frame with 200 observations on the following 5 variables. 
  \describe{
    \item{\code{X1}}{Covariate}
    \item{\code{X2}}{Covariate}
    \item{\code{male}}{Binary variable.}
    \item{\code{Z}}{Treatment indicator.}
    \item{\code{Y}}{Observed outcome.}    
  }
}

\keyword{datasets}
