\name{simData}
\alias{simData}
\docType{data}
\title{
Simulated data.
}
\description{
This dataset is used in several documentation files to illustrate the
different functionality of the package.
}
\usage{data("simData")}
\format{
  A data frame with 200 observations on the following 5 variables. 
  \describe{
    \item{\code{Y0}}{Potential outcome for the control.}
    \item{\code{Y1}}{Potential outcome for the treated.}
    \item{\code{Y}}{Observed outcome.}
    \item{\code{X}}{Covariate}
    \item{\code{Z}}{Treatment indicator.}
  }
}

\keyword{datasets}
